Prevent accidental directory traversal
Note: This is NOT a security vulnerability as uploading a web application effectively allows RCE anyway.

+++ java/org/apache/catalina/manager/ManagerServlet.java
@@ -713,6 +713,10 @@ public class ManagerServlet extends HttpServlet implements ContainerServlet {
         }
 
         File deployedWar = new File(host.getAppBaseFile(), baseName + ".war");
+        if (!pathCheck(deployedWar, host.getAppBaseFile(), writer, smClient)) {
+            // Any error reported in pathCheck()
+            return;
+        }
 
         // Determine full path for uploaded WAR
         File uploadedWar;
@@ -730,6 +734,10 @@ public class ManagerServlet extends HttpServlet implements ContainerServlet {
             }
         } else {
             File uploadPath = new File(versioned, tag);
+            if (!pathCheck(uploadPath, versioned, writer, smClient)) {
+                // Any error reported in pathCheck()
+                return;
+            }
             if (!uploadPath.mkdirs() && !uploadPath.isDirectory()) {
                 writer.println(smClient.getString("managerServlet.mkdirFail", uploadPath));
                 return;
@@ -815,8 +823,16 @@ public class ManagerServlet extends HttpServlet implements ContainerServlet {
 
         // Find the local WAR file
         File localWar = new File(new File(versioned, tag), baseName + ".war");
+        if (!pathCheck(localWar, versioned, writer, smClient)) {
+            // Any error reported in pathCheck()
+            return;
+        }
 
         File deployedWar = new File(host.getAppBaseFile(), baseName + ".war");
+        if (!pathCheck(deployedWar, host.getAppBaseFile(), writer, smClient)) {
+            // Any error reported in pathCheck()
+            return;
+        }
 
         // Copy WAR to appBase
         try {
@@ -847,6 +863,19 @@ public class ManagerServlet extends HttpServlet implements ContainerServlet {
         writeDeployResult(writer, smClient, name, displayPath);
     }
 
+
+    private static boolean pathCheck(File input, File expected, PrintWriter writer, StringManager smClient) {
+        try {
+            if (!input.getCanonicalFile().toPath().startsWith(expected.getCanonicalFile().toPath())) {
+                writer.println(smClient.getString("managerServlet.pathCheckFail", input, expected));
+                return false;
+            }
+        } catch (IOException ioe) {
+            writer.println(smClient.getString("managerServlet.pathCheckError", input, expected, ioe.getMessage()));
+            return false;
+        }
+        return true;
+    }
 
     /**
      * Install an application for the specified path from the specified web application archive.
@@ -916,7 +945,12 @@ public class ManagerServlet extends HttpServlet implements ContainerServlet {
                             return;
                         }
                         File localConfigFile = new File(configBase, baseName + ".xml");
+                        if (!pathCheck(localConfigFile, configBase, writer, smClient)) {
+                            // Any error reported in pathCheck()
+                            return;
+                        }
                         File configFile = new File(config);
+
                         // Skip delete and copy if source == destination
                         if (!configFile.getCanonicalPath().equals(localConfigFile.getCanonicalPath())) {
                             if (localConfigFile.isFile() && !localConfigFile.delete()) {
@@ -937,9 +971,18 @@ public class ManagerServlet extends HttpServlet implements ContainerServlet {
                         } else {
                             localWarFile = new File(host.getAppBaseFile(), baseName);
                         }
+                        if (!pathCheck(localWarFile, host.getAppBaseFile(), writer, smClient)) {
+                            // Any error reported in pathCheck()
+                            return;
+                        }
+
                         File warFile = new File(war);
                         if (!warFile.isAbsolute()) {
                             warFile = new File(host.getAppBaseFile(), war);
+                            if (!pathCheck(warFile, host.getAppBaseFile(), writer, smClient)) {
+                                // Any error reported in pathCheck()
+                                return;
+                            }
                         }
                         // Skip delete and copy if source == destination
                         if (!warFile.getCanonicalPath().equals(localWarFile.getCanonicalPath())) {
